Renamed profiles_plotter as plot_profiles to avoid name collision
